moving multistate samplers into new file, adding option for non linear lambda protocol
